fix wrong type in model

+++ backend/app/models.py
@@ -2,7 +2,7 @@ import enum
 from datetime import datetime
 from uuid import UUID, uuid4
 
-from sqlalchemy import JSON, Boolean, DateTime, Float, ForeignKey, String
+from sqlalchemy import JSON, Boolean, DateTime, Float, ForeignKey, Integer, String
 from sqlalchemy.orm import DeclarativeBase, Mapped, mapped_column
 
 
@@ -70,5 +70,5 @@ class Evaluation(Base):
     passed: Mapped[bool] = mapped_column(Boolean, nullable=False)
     reasoning: Mapped[JSON] = mapped_column(JSON, nullable=True)
     normalised_score: Mapped[float] = mapped_column(Float, nullable=False)
-    execution_time: Mapped[float] = mapped_column(Float, nullable=True)
+    execution_time: Mapped[int] = mapped_column(Integer, nullable=True)
     error: Mapped[str] = mapped_column(String, nullable=True)
